Using LifecycleManager to handle SDK lifecycle for all components. Fixed broken tests due to refactor
