+++ src/androidTest/java/fake/ImpressionsManagerStub.java
@@ -1,0 +1,25 @@
+package fake;
+
+import io.split.android.client.impressions.ImpressionsManager;
+
+public class ImpressionsManagerStub implements ImpressionsManager {
+    @Override
+    public void pause() {
+
+    }
+
+    @Override
+    public void resume() {
+
+    }
+
+    @Override
+    public void flushImpressions() {
+
+    }
+
+    @Override
+    public void saveToDisk() {
+
+    }
+}
