+++ src/androidTest/java/fake/RefreshableMySegmentsFetcherProviderStub.java
@@ -1,22 +1,14 @@
-package io.split.android.fake;
+package fake;
 
 import java.io.IOException;
-import java.util.List;
 
 import io.split.android.engine.segments.MySegments;
 import io.split.android.engine.segments.RefreshableMySegmentsFetcherProvider;
 
 public class RefreshableMySegmentsFetcherProviderStub implements RefreshableMySegmentsFetcherProvider {
-
-    MySegments mMySegments;
-
-    public RefreshableMySegmentsFetcherProviderStub(List<String> mySegments) {
-        this.mMySegments = new MySegmentsStub(mySegments);
-    }
-
     @Override
     public MySegments mySegments() {
-        return mMySegments;
+        return null;
     }
 
     @Override
@@ -31,5 +23,6 @@ public class RefreshableMySegmentsFetcherProviderStub implements RefreshableMySe
 
     @Override
     public void close() throws IOException {
+
     }
 }
