+++ src/main/java/io/split/android/client/SplitFactoryImpl.java
@@ -12,8 +12,10 @@ import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 
 import io.split.android.client.api.Key;
+import io.split.android.client.cache.IMySegmentsCache;
 import io.split.android.client.cache.ISplitCache;
 import io.split.android.client.cache.ISplitChangeCache;
+import io.split.android.client.cache.MySegmentsCache;
 import io.split.android.client.cache.SplitCache;
 import io.split.android.client.cache.SplitChangeCache;
 import io.split.android.client.events.SplitEventsManager;
@@ -22,7 +24,7 @@ import io.split.android.client.factory.FactoryMonitorImpl;
 import io.split.android.client.impressions.IImpressionsStorage;
 import io.split.android.client.impressions.ImpressionListener;
 import io.split.android.client.impressions.ImpressionsFileStorage;
-import io.split.android.client.impressions.ImpressionsManager;
+import io.split.android.client.impressions.ImpressionsManagerImpl;
 import io.split.android.client.impressions.ImpressionsStorageManager;
 import io.split.android.client.impressions.ImpressionsStorageManagerConfig;
 import io.split.android.client.lifecycle.LifecycleManager;
@@ -48,6 +50,7 @@ import io.split.android.client.validators.ValidationMessageLogger;
 import io.split.android.client.validators.ValidationMessageLoggerImpl;
 import io.split.android.engine.SDKReadinessGates;
 import io.split.android.engine.experiments.RefreshableSplitFetcherProvider;
+import io.split.android.engine.experiments.RefreshableSplitFetcherProviderImpl;
 import io.split.android.engine.experiments.SplitChangeFetcher;
 import io.split.android.engine.experiments.SplitParser;
 import io.split.android.engine.segments.MySegmentsFetcher;
@@ -69,8 +72,8 @@ public class SplitFactoryImpl implements SplitFactory {
     private SDKReadinessGates gates;
 
     private TrackClient _trackClient;
-    FactoryMonitor _factoryMonitor = FactoryMonitorImpl.getSharedInstance();
-    LifecycleManager _lifecyleManager;
+    private FactoryMonitor _factoryMonitor = FactoryMonitorImpl.getSharedInstance();
+    private LifecycleManager _lifecyleManager;
 
     public SplitFactoryImpl(String apiToken, Key key, SplitClientConfig config, Context context) throws IOException, InterruptedException, TimeoutException, URISyntaxException {
 
@@ -123,7 +126,8 @@ public class SplitFactoryImpl implements SplitFactory {
 
         // Segments
         IStorage mySegmentsStorage = new FileStorage(context.getCacheDir(), dataFolderName);
-        MySegmentsFetcher mySegmentsFetcher = HttpMySegmentsFetcher.create(httpClient, rootTarget, mySegmentsStorage);
+        IMySegmentsCache mySegmentsCache = new MySegmentsCache(mySegmentsStorage);
+        MySegmentsFetcher mySegmentsFetcher = HttpMySegmentsFetcher.create(httpClient, rootTarget, mySegmentsCache);
         final RefreshableMySegmentsFetcherProviderImpl segmentFetcher = new RefreshableMySegmentsFetcherProviderImpl(mySegmentsFetcher, findPollingPeriod(RANDOM, config.segmentsRefreshRate()), key.matchingKey(), _eventsManager);
 
         SplitParser splitParser = new SplitParser(segmentFetcher);
@@ -134,7 +138,7 @@ public class SplitFactoryImpl implements SplitFactory {
         ISplitChangeCache splitChangeCache = new SplitChangeCache(splitCache);
 
         SplitChangeFetcher splitChangeFetcher = HttpSplitChangeFetcher.create(httpClient, rootTarget, uncachedFireAndForget, splitChangeCache);
-        final RefreshableSplitFetcherProvider splitFetcherProvider = new RefreshableSplitFetcherProvider(
+        final RefreshableSplitFetcherProvider splitFetcherProvider = new RefreshableSplitFetcherProviderImpl(
                 splitChangeFetcher, splitParser, findPollingPeriod(RANDOM,
                 config.featuresRefreshRate()), _eventsManager, splitCache.getChangeNumber());
 
@@ -144,7 +148,7 @@ public class SplitFactoryImpl implements SplitFactory {
         impressionsStorageManagerConfig.setImpressionsChunkOudatedTime(config.impressionsChunkOutdatedTime());
         IImpressionsStorage impressionsStorage = new ImpressionsFileStorage(context.getCacheDir(), dataFolderName);
         final ImpressionsStorageManager impressionsStorageManager = new ImpressionsStorageManager(impressionsStorage, impressionsStorageManagerConfig);
-        final ImpressionsManager splitImpressionListener = ImpressionsManager.instance(httpClient, config, impressionsStorageManager);
+        final ImpressionsManagerImpl splitImpressionListener = ImpressionsManagerImpl.instance(httpClient, config, impressionsStorageManager);
         final ImpressionListener impressionListener;
 
         if (config.impressionListener() != null) {
@@ -172,12 +176,14 @@ public class SplitFactoryImpl implements SplitFactory {
         _trackClient = TrackClientImpl.create(trackConfig, httpClient, eventsRootTarget, trackStorageManager, splitCache);
 
 
-        _lifecyleManager = new LifecycleManager(splitImpressionListener, _trackClient, splitFetcherProvider, segmentFetcher);
+        _lifecyleManager = new LifecycleManager(splitImpressionListener, _trackClient, splitFetcherProvider, segmentFetcher, splitCache, mySegmentsCache);
 
         destroyer = new Runnable() {
             public void run() {
                 Logger.w("Shutdown called for split");
                 try {
+                    _lifecyleManager.destroy();
+                    Logger.i("Successful shutdown of lifecycle manager");
                     _factoryMonitor.remove(_apiKey);
                     _trackClient.close();
                     Logger.i("Successful shutdown of Track client");
@@ -195,8 +201,6 @@ public class SplitFactoryImpl implements SplitFactory {
                     Logger.i("Successful shutdown of httpclient");
                     _manager.destroy();
                     Logger.i("Successful shutdown of manager");
-                    _lifecyleManager.destroy();
-                    Logger.i("Successful shutdown of lifecycle manager");
 
                 } catch (Exception e) {
                     Logger.e(e, "We could not shutdown split");
