+++ src/main/java/io/split/android/client/TrackClient.java
@@ -13,4 +13,6 @@ public interface TrackClient {
     void close();
 
     void flush();
+
+    void saveToDisk();
 }
