+++ src/main/java/io/split/android/client/cache/IMySegmentsCache.java
@@ -29,4 +29,9 @@ public interface IMySegmentsCache {
      * Deletes the list of MySegments from the cache
      */
     void deleteMySegments(String key);
+
+    /**
+     * Saves in memory cash to disk
+     */
+     void saveToDisk();
 }
