+++ src/main/java/io/split/android/client/cache/ISplitCache.java
@@ -47,4 +47,9 @@ public interface ISplitCache {
      * @return true if it exists, false otherwise
      */
     boolean trafficTypeExists(String trafficType);
+
+    /**
+     * Saves in memory cash to disk
+     */
+    void saveToDisk();
 }
