+++ src/main/java/io/split/android/client/lifecycle/LifecycleManager.java
@@ -6,6 +6,8 @@ import android.arch.lifecycle.OnLifecycleEvent;
 import android.arch.lifecycle.ProcessLifecycleOwner;
 
 import io.split.android.client.TrackClient;
+import io.split.android.client.cache.IMySegmentsCache;
+import io.split.android.client.cache.ISplitCache;
 import io.split.android.client.impressions.ImpressionsManager;
 import io.split.android.engine.experiments.RefreshableSplitFetcherProvider;
 import io.split.android.engine.segments.RefreshableMySegmentsFetcherProvider;
@@ -16,16 +18,23 @@ public class LifecycleManager implements LifecycleObserver {
     TrackClient mTrackClient;
     RefreshableSplitFetcherProvider mSplitFetcherProvider;
     RefreshableMySegmentsFetcherProvider mMySegmentsFetcherProvider;
+    IMySegmentsCache mMySegmentsCache;
+    ISplitCache mSplitCache;
 
     public LifecycleManager(ImpressionsManager impressionsManager,
                             TrackClient trackClient,
                             RefreshableSplitFetcherProvider splitFetcherProvider,
-                            RefreshableMySegmentsFetcherProvider mySegmentsFetcherProvider) {
+                            RefreshableMySegmentsFetcherProvider mySegmentsFetcherProvider,
+                            ISplitCache splitCache,
+                            IMySegmentsCache mySegmentsCache) {
 
         mImpressionsManager = impressionsManager;
         mTrackClient = trackClient;
         mSplitFetcherProvider = splitFetcherProvider;
         mMySegmentsFetcherProvider = mySegmentsFetcherProvider;
+        mSplitCache = splitCache;
+        mMySegmentsCache = mySegmentsCache;
+
         ProcessLifecycleOwner.get().getLifecycle().addObserver(this);
     }
 
@@ -33,10 +42,12 @@ public class LifecycleManager implements LifecycleObserver {
     private void onPause() {
         if(mImpressionsManager != null) {
             mImpressionsManager.pause();
+            mImpressionsManager.saveToDisk();
         }
 
         if(mTrackClient != null) {
             mTrackClient.pause();
+            mTrackClient.saveToDisk();
         }
 
         if(mSplitFetcherProvider != null) {
@@ -45,6 +56,14 @@ public class LifecycleManager implements LifecycleObserver {
 
         if(mMySegmentsFetcherProvider != null) {
             mMySegmentsFetcherProvider.pause();
+        }
+
+        if(mSplitCache != null) {
+            mSplitCache.saveToDisk();
+        }
+
+        if(mMySegmentsCache != null) {
+            mMySegmentsCache.saveToDisk();
         }
     }
