+++ src/test/java/io/split/android/client/impressions/ImpressionsManagerTest.java
@@ -14,7 +14,6 @@ import java.util.List;
 import io.split.android.client.SplitClientConfig;
 import io.split.android.client.dtos.KeyImpression;
 import io.split.android.client.dtos.TestImpressions;
-import io.split.android.client.storage.MemoryStorage;
 import io.split.android.fake.ImpressionsFileStorageStub;
 
 import static org.hamcrest.Matchers.equalTo;
@@ -41,7 +40,7 @@ public class ImpressionsManagerTest {
         ImpressionsSender senderMock = Mockito.mock(ImpressionsSender.class);
         ImpressionsStorageManager storageMock = new ImpressionsStorageManager(new ImpressionsFileStorageStub(), storageConfig());
 
-        ImpressionsManager treatmentLog = ImpressionsManager.instanceForTest(null, config, senderMock, storageMock);
+        ImpressionsManagerImpl treatmentLog = ImpressionsManagerImpl.instanceForTest(null, config, senderMock, storageMock);
 
         KeyImpression ki1 = keyImpression("test1", "adil", "on", 1L, null);
         KeyImpression ki2 = keyImpression("test1", "adil", "on", 2L, 1L);
@@ -75,7 +74,7 @@ public class ImpressionsManagerTest {
         ImpressionsSender senderMock = Mockito.mock(ImpressionsSender.class);
         ImpressionsStorageManager storageMock = new ImpressionsStorageManager(new ImpressionsFileStorageStub(), storageConfig());
 
-        ImpressionsManager treatmentLog = ImpressionsManager.instanceForTest(null, config, senderMock, storageMock);
+        ImpressionsManagerImpl treatmentLog = ImpressionsManagerImpl.instanceForTest(null, config, senderMock, storageMock);
 
         // These 4 unique test name will cause 4 entries but we are caping at the first 3.
         KeyImpression ki1 = keyImpression("test1", "adil", "on", 1L, null);
@@ -110,7 +109,7 @@ public class ImpressionsManagerTest {
         ImpressionsSender senderMock = Mockito.mock(ImpressionsSender.class);
         ImpressionsStorageManager storageMock = new ImpressionsStorageManager(new ImpressionsFileStorageStub(), storageConfig());
 
-        ImpressionsManager treatmentLog = ImpressionsManager.instanceForTest(null, config, senderMock, storageMock);
+        ImpressionsManagerImpl treatmentLog = ImpressionsManagerImpl.instanceForTest(null, config, senderMock, storageMock);
 
         // These 4 unique test name will cause 4 entries but we are caping at the first 3.
         KeyImpression ki1 = keyImpression("test1", "adil", "on", 1L, 1L);
@@ -147,7 +146,7 @@ public class ImpressionsManagerTest {
         ImpressionsSender senderMock = Mockito.mock(ImpressionsSender.class);
         ImpressionsStorageManager storageMock = new ImpressionsStorageManager(new ImpressionsFileStorageStub(), storageConfig());
 
-        ImpressionsManager treatmentLog = ImpressionsManager.instanceForTest(null, config, senderMock, storageMock);
+        ImpressionsManagerImpl treatmentLog = ImpressionsManagerImpl.instanceForTest(null, config, senderMock, storageMock);
 
         // There are no impressions to post.
